fix: add gunshi minimal entry point
