+++ packages/gunshi/src/index.ts
@@ -1,5 +1,13 @@
 /**
- * The main entry point for Gunshi.
+ * gunshi cli entry point.
+ *
+ * This entry point exports the bellow APIs and types:
+ * - `cli`: The main CLI function to run the command, included `global` and `renderer` built-in plugins.
+ * - `define`: A function to define a command.
+ * - `lazy`: A function to lazily load a command.
+ * - `plugin`: A function to create a plugin.
+ * - `args-tokens` utilities, `parseArgs` and `resolveArgs` for parsing command line arguments.
+ * - Some basic type definitions, such as `CommandContext`, `Plugin`, `PluginContext`, etc.
  *
  * @example
  * ```js
